Create new component and move there header part from PanelComponent, add spec
