+++ app/components/avo/panel_component.html.erb
@@ -1,38 +1,17 @@
 <%= content_tag :div, class: classes, data: data_attributes do %>
   <%= render Avo::CoverPhotoComponent.new cover_photo: @cover_photo %>
-
-  <% if render_header? %>
-    <div data-target="panel-header" class="flex flex-col w-full mb-4">
-      <div class="flex justify-center sm:justify-start flex-col sm:flex-row w-full flex-1 has-cover-photo:mt-0">
-        <%= render Avo::ProfilePhotoComponent.new profile_photo: @profile_photo %>
-        <div class="flex flex-col flex-1 w-full">
-          <% if display_breadcrumbs? %>
-            <div class="breadcrumbs text-center sm:text-left mb-2">
-              <%= helpers.render_avo_breadcrumbs(separator: helpers.svg("chevron-right", class: "inline-block h-3 stroke-current relative top-[-1px] ml-1")) if Avo.configuration.display_breadcrumbs %>
-            </div>
-          <% end %>
-          <div class="flex-1 flex flex-col xl:flex-row justify-between gap-1 grow-0">
-            <div class="overflow-hidden flex flex-col">
-              <% if name_slot? %>
-                <%= name_slot %>
-              <% else %>
-                <%= render Avo::PanelNameComponent.new name: @name %>
-              <% end %>
-              <% if description.present? %>
-                <div class="text-sm tracking-normal font-medium text-gray-600 text-center sm:text-left" data-target="description">
-                  <%== description %>
-                </div>
-              <% end %>
-            </div>
-            <% if tools.present? %>
-              <div class="flex-1 w-full flex flex-col sm:flex-row xl:justify-end sm:items-end gap-2 mt-4 xl:mt-0" data-target="panel-tools">
-                <%= tools %>
-              </div>
-            <% end %>
-          </div>
-        </div>
-      </div>
-    </div>
+  <%= render Avo::PanelHeaderComponent.new(
+    name: @name,
+    description: @description,
+    display_breadcrumbs: @display_breadcrumbs,
+    profile_photo: @profile_photo
+  ) do |header| %>
+    <% header.name_slot do %>
+      <%= name_slot %>
+    <% end %>
+    <% header.tools do %>
+      <%= tools %>
+    <% end %>
   <% end %>
   <% if body? %>
     <div data-target="panel-body" class="flex flex-col sm:flex-row space-y-4 sm:space-y-0 sm:gap-4 w-full <% if sidebar? %> has-sidebar <% end %>">
