+++ spec/components/avo/panel_header_component_spec.rb
@@ -1,0 +1,11 @@
+require "rails_helper"
+
+RSpec.describe Avo::PanelHeaderComponent, type: :component do
+  # it "renders something useful" do
+  #   expect(
+  #     render_inline(described_class.new(attr: "value")) { "Hello, components!" }.css("p").to_html
+  #   ).to include(
+  #     "Hello, components!"
+  #   )
+  # end
+end
